Add replob v5

+++ src/stats.h
@@ -16,13 +16,13 @@
 
 struct GlobalStats
 {
-    int iterations = 0;
-    int disconnects = 0;
+    int64_t iterations = 0;
+    int64_t disconnects = 0;
 };
 
 struct Stats : Initer<Stats>
 {
-    int disconnects = 0;
+    int64_t disconnects = 0;
 };
 
 inline std::ostream& operator<<(std::ostream& o, const GlobalStats& s)
